+++ readium/navigator/src/main/java/org/readium/r2/navigator/epub/EpubNavigatorFragment.kt
@@ -265,6 +265,19 @@ public class EpubNavigatorFragment internal constructor(
 
     public interface Listener : OverflowableNavigator.Listener, HyperlinkNavigator.Listener
 
+    private sealed class State {
+        /** The navigator just started and didn't load any resource yet. */
+        data object Initializing : State()
+
+        /** The navigator is loading the first resource at `initialResourceHref`. */
+        data class Loading(val initialResourceHref: Url) : State()
+
+        /** The navigator is fully initialized and ready for action. */
+        data object Ready : State()
+    }
+
+    private var state: State = State.Initializing
+
     // Configurable
 
     override val settings: StateFlow<EpubSettings> get() = viewModel.settings
@@ -591,9 +604,13 @@ public class EpubNavigatorFragment internal constructor(
     }
 
     @OptIn(DelicateReadiumApi::class)
-    override fun go(locator: Locator, animated: Boolean, completion: () -> Unit): Boolean {
+    override fun go(locator: Locator, animated: Boolean): Boolean {
         @Suppress("NAME_SHADOWING")
         val locator = publication.normalizeLocator(locator)
+
+        if (state == State.Initializing) {
+            state = State.Loading(locator.href)
+        }
 
         listener?.onJumpToLocator(locator)
 
@@ -636,9 +653,9 @@ public class EpubNavigatorFragment internal constructor(
         return true
     }
 
-    override fun go(link: Link, animated: Boolean, completion: () -> Unit): Boolean {
+    override fun go(link: Link, animated: Boolean): Boolean {
         val locator = publication.locatorFromLink(link) ?: return false
-        return go(locator, animated, completion)
+        return go(locator, animated)
     }
 
     private fun run(commands: List<RunScriptCommand>) {
@@ -759,12 +776,18 @@ public class EpubNavigatorFragment internal constructor(
         override val readingProgression: ReadingProgression
             get() = viewModel.readingProgression
 
-        override fun onResourceLoaded(link: Link?, webView: R2BasicWebView, url: String?) {
-            run(viewModel.onResourceLoaded(link, webView))
+        override fun onResourceLoaded(webView: R2BasicWebView, link: Link) {
+            run(viewModel.onResourceLoaded(webView, link))
         }
 
-        override fun onPageLoaded() {
+        override fun onPageLoaded(webView: R2BasicWebView, link: Link) {
             paginationListener?.onPageLoaded()
+
+            val href = link.url()
+            if (state is State.Initializing || (state as? State.Loading)?.initialResourceHref == href) {
+                state = State.Ready
+            }
+
             notifyCurrentLocation()
         }
 
@@ -846,9 +869,9 @@ public class EpubNavigatorFragment internal constructor(
                 ?.let { publication.get(it) }
     }
 
-    override fun goForward(animated: Boolean, completion: () -> Unit): Boolean {
+    override fun goForward(animated: Boolean): Boolean {
         if (publication.metadata.presentation.layout == EpubLayout.FIXED) {
-            return goToNextResource(jump = false, animated = animated, completion)
+            return goToNextResource(jump = false, animated = animated)
         }
 
         val webView = currentReflowablePageFragment?.webView ?: return false
@@ -860,13 +883,12 @@ public class EpubNavigatorFragment internal constructor(
             ReadingProgression.RTL ->
                 webView.scrollLeft(animated)
         }
-        lifecycleScope.launch { completion() }
         return true
     }
 
-    override fun goBackward(animated: Boolean, completion: () -> Unit): Boolean {
+    override fun goBackward(animated: Boolean): Boolean {
         if (publication.metadata.presentation.layout == EpubLayout.FIXED) {
-            return goToPreviousResource(jump = false, animated = animated, completion)
+            return goToPreviousResource(jump = false, animated = animated)
         }
 
         val webView = currentReflowablePageFragment?.webView ?: return false
@@ -878,11 +900,10 @@ public class EpubNavigatorFragment internal constructor(
             ReadingProgression.RTL ->
                 webView.scrollRight(animated)
         }
-        lifecycleScope.launch { completion() }
         return true
     }
 
-    private fun goToNextResource(jump: Boolean, animated: Boolean, completion: () -> Unit = {}): Boolean {
+    private fun goToNextResource(jump: Boolean, animated: Boolean): Boolean {
         val adapter = resourcePager.adapter ?: return false
         if (resourcePager.currentItem >= adapter.count - 1) {
             return false
@@ -902,11 +923,10 @@ public class EpubNavigatorFragment internal constructor(
             }
         }
 
-        viewLifecycleOwner.lifecycleScope.launch { completion() }
         return true
     }
 
-    private fun goToPreviousResource(jump: Boolean, animated: Boolean, completion: () -> Unit = {}): Boolean {
+    private fun goToPreviousResource(jump: Boolean, animated: Boolean): Boolean {
         if (resourcePager.currentItem <= 0) {
             return false
         }
@@ -925,7 +945,6 @@ public class EpubNavigatorFragment internal constructor(
             }
         }
 
-        viewLifecycleOwner.lifecycleScope.launch { completion() }
         return true
     }
 
@@ -1032,7 +1051,9 @@ public class EpubNavigatorFragment internal constructor(
         debounceLocationNotificationJob = viewLifecycleOwner.lifecycleScope.launch {
             delay(100L)
 
-            if (currentReflowablePageFragment?.isLoaded?.value == false) {
+            // We don't want to notify the current location if the navigator is still loading a
+            // locator, to avoid notifying intermediate locations.
+            if (currentReflowablePageFragment?.isLoaded?.value == false || state != State.Ready) {
                 return@launch
             }