+++ readium/shared/src/main/java/org/readium/r2/shared/extensions/String.kt
@@ -12,30 +12,26 @@ package org.readium.r2.shared.extensions
 import android.net.Uri
 import java.net.URL
 import java.security.MessageDigest
-import java.util.*
-import org.joda.time.DateTime
-import org.joda.time.DateTimeZone
+import java.util.Date
+import kotlinx.datetime.Instant
+import kotlinx.datetime.LocalDate
+import kotlinx.datetime.LocalDateTime
+import kotlinx.datetime.TimeZone
+import kotlinx.datetime.atStartOfDayIn
+import kotlinx.datetime.toInstant
 import org.json.JSONException
 import org.json.JSONObject
 import org.readium.r2.shared.InternalReadiumApi
 
 @InternalReadiumApi
-public fun String.iso8601ToDate(): Date? =
-    try {
-        // We assume that a date without a time zone component is in UTC. To handle this properly,
-        // we need to set the default time zone of Joda to UTC, since by default it uses the local
-        // time zone. This ensures that apps see exactly the same dates (e.g. published) no matter
-        // where they are located.
-        // For the same reason, the output Date will be in UTC. Apps should convert it to the local
-        // time zone for display purposes, or keep it as UTC for storage.
-        val defaultTZ = DateTimeZone.getDefault()
-        DateTimeZone.setDefault(DateTimeZone.UTC)
-        val date = DateTime(this).toDateTime(DateTimeZone.UTC).toDate()
-        DateTimeZone.setDefault(defaultTZ)
-        date
-    } catch (e: Exception) {
-        null
-    }
+public fun String.iso8601ToDate(): Date? {
+    val instant = tryOrNull { Instant.parse(this) }
+        ?: tryOrNull { LocalDateTime.parse(this).toInstant(TimeZone.UTC) }
+        ?: tryOrNull { LocalDate.parse(this).atStartOfDayIn(TimeZone.UTC) }
+        ?: return null
+
+    return Date(instant.toEpochMilliseconds())
+}
 
 /**
  * If this string starts with the given [prefix], returns this string.