+++ readium/shared/src/main/java/org/readium/r2/shared/util/zip/FileZipArchiveProvider.kt
@@ -19,7 +19,7 @@ import org.readium.r2.shared.util.data.Container
 import org.readium.r2.shared.util.data.ReadError
 import org.readium.r2.shared.util.file.FileSystemError
 import org.readium.r2.shared.util.format.Format
-import org.readium.r2.shared.util.format.ZipSpecification
+import org.readium.r2.shared.util.format.Specification
 import org.readium.r2.shared.util.resource.Resource
 
 /**
@@ -54,7 +54,7 @@ internal class FileZipArchiveProvider {
         format: Format,
         file: File
     ): Try<Container<Resource>, ArchiveOpener.OpenError> {
-        if (!format.conformsTo(ZipSpecification)) {
+        if (!format.conformsTo(Specification.Zip)) {
             return Try.failure(
                 ArchiveOpener.OpenError.FormatNotSupported(format)
             )